+++ src/client.ts
@@ -34,7 +34,7 @@ export type OnReadyResult = {
 };
 
 const REACT_SDK_CLIENT_ENGINE = 'react-sdk';
-const REACT_SDK_CLIENT_VERSION = '1.2.0-alpha.1';
+const REACT_SDK_CLIENT_VERSION = '2.0.0-alpha.1';
 
 export interface ReactSDKClient extends optimizely.Client {
   user: UserContext;